refactor: add table size variables and rewrite board definition
The previous implementation tied the board construction to 3 by 3
boards. In order to generalize the gameBoard module two new variables
were added and the board default construction was rewrited.

now the row and column counts are available to other functions inside
the module, avoiding this way to write them with 'magic numbers'
as the board size.